fix(store/selectors): Fix getTotalDurationByDomain domain grouping
Previously we remove "https://"/"www." strings from the domain after
they were by grouped & computed, but this doesn't give us the correct
result as we want to add up visits to a domain regardless of whether
"https://" or "www." is present.

This change modifies the selector to clean up the domain before the
computation step.